+++ backend/jobplatform/src/main/java/jobplatform/fo/enterprise/domain/dto/ResumeSearchDataDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ResumeSearchDataDTO {
 	// 이력서 번호
-	private int mbr_sq;
+	private Long mbr_sq;
 	//정렬
 	private String sort;
 	// 현재 페이지 번호
